requirements.txt
pypdf2
pymupdf
fitz
camelot-py
opencv-python